+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- **[NEW]** Add `LockIcon`
+- **[UPDATE]** Fix `Button` with `Icon` display case
 - [...]
 
 # v22.0.0 (26/02/2020)